Modify host header when proxying to public ui
Where API services have been made public then the host header for the proxied request needs to match that configured on the ingress as the url for the service. Otherwise the proxied request is redirected back to the source service on account of its host header, and it throws into an infinite loop.