Use cached data for establishment licence fees table
The number of PILs in prod (~20,000) makes the billing query across all establishments quite slow, and so rendering the table currently times out occasionally.

We already keep a copy of all the data in the table in memory as part of the cache to serve the total fees data. We can use this to render the table without having to go to the database to do a sort and filter.

This means we should be able to serve the page reliably in a fraction of a second from the cached data.